+++ lib/shared/src/context/openctx/api.ts
@@ -1,39 +1,23 @@
-import type { Annotation, ItemsParams, ItemsResult } from '@openctx/client'
-import type { TextDocument } from 'vscode'
-import type { RangeData } from '../../common/range'
+import type { Client } from '@openctx/client'
+import type * as vscode from 'vscode'
+
+type OpenCtxClient = Client<vscode.Range>
+
+let _client: OpenCtxClient | undefined
 
 /**
- * Copied from OpenCtx's VS Code extension sources.
+ * Set the handle to the OpenCtx client.
  */
-export interface OpenCtxExtensionAPI {
-    getItems(params: ItemsParams): Promise<ItemsResult | null>
-    getAnnotations(doc: Pick<TextDocument, 'uri' | 'getText'>): Promise<Annotation<RangeData>[] | null>
-}
-
-let _getAPI: (() => Promise<OpenCtxExtensionAPI | null>) | undefined
-
-/**
- * Set the handle to the OpenCtx extension API (e.g., from
- * `vscode.extensions.getExtension('sourcegraph.openctx')`).
- */
-export function setOpenCtxExtensionAPI(getAPI: () => Promise<OpenCtxExtensionAPI | null>): void {
-    if (_getAPI) {
+export function setOpenCtxClient(client: OpenCtxClient | undefined): void {
+    if (_client) {
         throw new Error('OpenCtx extension API is already set')
     }
-    _getAPI = getAPI
+    _client = client
 }
 
-let _api: Promise<OpenCtxExtensionAPI | null> | undefined
-
 /**
- * Get a handle to the OpenCtx extension API, set in {@link setOpenCtxExtensionAPI}.
+ * Get a handle to the OpenCtx client, set in {@link setOpenCtxClient}.
  */
-export function getOpenCtxExtensionAPI(): Promise<OpenCtxExtensionAPI | null | undefined> {
-    if (!_getAPI) {
-        return Promise.resolve(undefined)
-    }
-    if (!_api) {
-        _api = _getAPI()
-    }
-    return _api
+export function getOpenCtxClient(): OpenCtxClient | undefined {
+    return _client
 }