+++ vscode/src/chat/context/chatContext.ts
@@ -3,8 +3,11 @@ import {
     type ContextItem,
     type ContextItemProps,
     type ContextMentionProvider,
+    FILE_CONTEXT_MENTION_PROVIDER,
     type MentionQuery,
-    parseMentionQuery,
+    PACKAGE_CONTEXT_MENTION_PROVIDER,
+    SYMBOL_CONTEXT_MENTION_PROVIDER,
+    URL_CONTEXT_MENTION_PROVIDER,
 } from '@sourcegraph/cody-shared'
 import * as vscode from 'vscode'
 import { getContextFileFromUri } from '../../commands/context/file-path'
@@ -16,20 +19,17 @@ import {
 } from '../../editor/utils/editor-context'
 
 export async function getChatContextItemsForMention(
-    query: MentionQuery | string,
+    mentionQuery: MentionQuery,
     cancellationToken: vscode.CancellationToken,
     telemetryRecorder?: {
         empty: () => void
         withProvider: (type: MentionQuery['provider']) => void
     }
 ): Promise<ContextItem[]> {
-    const mentionQuery =
-        typeof query === 'string' ? parseMentionQuery(query, getEnabledContextMentionProviders()) : query
-
     // Logging: log when the at-mention starts, and then log when we know the type (after the 1st
     // character is typed). Don't log otherwise because we would be logging prefixes of the same
     // query repeatedly, which is not needed.
-    if (mentionQuery.provider === 'default') {
+    if (mentionQuery.provider === null) {
         telemetryRecorder?.empty()
     } else {
         telemetryRecorder?.withProvider(mentionQuery.provider)
@@ -37,13 +37,15 @@ export async function getChatContextItemsForMention(
 
     const MAX_RESULTS = 20
     switch (mentionQuery.provider) {
-        case 'default':
+        case null:
             return getOpenTabsContextFile()
-        case 'symbol':
+        case SYMBOL_CONTEXT_MENTION_PROVIDER.id:
             // It would be nice if the VS Code symbols API supports cancellation, but it doesn't
             return getSymbolContextFiles(mentionQuery.text, MAX_RESULTS)
-        case 'file': {
-            const files = await getFileContextFiles(mentionQuery.text, MAX_RESULTS)
+        case FILE_CONTEXT_MENTION_PROVIDER.id: {
+            const files = mentionQuery.text
+                ? await getFileContextFiles(mentionQuery.text, MAX_RESULTS)
+                : await getOpenTabsContextFile()
 
             // If a range is provided, that means user is trying to mention a specific line range.
             // We will get the content of the file for that range to display file size warning if needed.
@@ -89,8 +91,8 @@ export function getEnabledContextMentionProviders(): ContextMentionProvider[] {
     if (isURLProviderEnabled || isPackageProviderEnabled) {
         return CONTEXT_MENTION_PROVIDERS.filter(
             provider =>
-                (isURLProviderEnabled && provider.id === 'url') ||
-                (isPackageProviderEnabled && provider.id === 'package')
+                (isURLProviderEnabled && provider.id === URL_CONTEXT_MENTION_PROVIDER.id) ||
+                (isPackageProviderEnabled && provider.id === PACKAGE_CONTEXT_MENTION_PROVIDER.id)
         )
     }
     return []