+++ vscode/src/configuration.test.ts
@@ -104,6 +104,10 @@ describe('getConfiguration', () => {
                         return false
                     case 'cody.experimental.supercompletions':
                         return false
+                    case 'cody.experimental.noodle':
+                        return false
+                    case 'cody.experimental.urlContext':
+                        return false
                     case 'cody.experimental.github.accessToken':
                         return ''
                     default:
@@ -133,6 +137,8 @@ describe('getConfiguration', () => {
             experimentalGuardrails: true,
             experimentalOllamaChat: true,
             experimentalCommitMessage: true,
+            experimentalNoodle: false,
+            experimentalURLContext: false,
             experimentalGithubAccessToken: '',
             codeActions: true,
             commandHints: true,