+++ vscode/src/services/UpstreamHealthProvider.ts
@@ -3,7 +3,9 @@ import {
     type ConfigurationWithAccessToken,
     addCustomUserAgent,
     addTraceparent,
+    isDotCom,
     logDebug,
+    wrapInActiveSpan,
 } from '@sourcegraph/cody-shared'
 import { fetch } from '@sourcegraph/cody-shared'
 import type * as vscode from 'vscode'
@@ -11,7 +13,7 @@ import type * as vscode from 'vscode'
 // We choose an interval that gives us a reasonable aggregate without causing
 // too many requests
 const PING_INTERVAL_MS = 10 * 60 * 1000 // 10 minutes
-const TWO_HOURS_MS = 2 * 60 * 60 * 1000
+const INITIAL_PING_DELAY_MS = 10 * 1000 // 10 seconds
 
 /**
  * A provider that regularly pings the connected Sourcegraph instance to
@@ -20,8 +22,8 @@ const TWO_HOURS_MS = 2 * 60 * 60 * 1000
  * You can query it to get aggregates of the most recent pings.
  */
 export class UpstreamHealthProvider implements vscode.Disposable {
-    // Array sorted by duration for easy median calculation
-    private recentDurationsSorted: { timestamp: number; duration: number }[] = []
+    private lastUpstreamLatency?: number
+    private lastGatewayLatency?: number
 
     private config: Pick<
         ConfigurationWithAccessToken,
@@ -29,22 +31,35 @@ export class UpstreamHealthProvider implements vscode.Disposable {
     > | null = null
     private nextTimeoutId: NodeJS.Timeout | null = null
 
-    public getMedianDuration(): number | undefined {
+    public getUpstreamLatency(): number | undefined {
         if (!this.config) {
             return undefined
         }
-        if (this.recentDurationsSorted.length === 0) {
+        return this.lastUpstreamLatency
+    }
+
+    public getGatewayLatency(): number | undefined {
+        if (!this.config) {
             return undefined
         }
-        return this.recentDurationsSorted[Math.floor(this.recentDurationsSorted.length / 2)].duration
+        return this.lastGatewayLatency
     }
 
     public onConfigurationChange(
         newConfig: Pick<ConfigurationWithAccessToken, 'serverEndpoint' | 'customHeaders' | 'accessToken'>
     ): this {
         this.config = newConfig
-        this.recentDurationsSorted = []
-        this.measure()
+        this.lastUpstreamLatency = undefined
+        this.lastGatewayLatency = undefined
+
+        // Enqueue the initial ping after a config change in 10 seconds. This
+        // avoids running the test while the extension is still initializing and
+        // competing with many other network requests.
+        if (this.nextTimeoutId) {
+            clearTimeout(this.nextTimeoutId)
+        }
+        this.nextTimeoutId = setTimeout(this.measure.bind(this), INITIAL_PING_DELAY_MS)
+
         return this
     }
 
@@ -53,49 +68,71 @@ export class UpstreamHealthProvider implements vscode.Disposable {
             clearTimeout(this.nextTimeoutId)
         }
 
-        const start = Date.now()
         try {
+            if (process.env.DISABLE_UPSTREAM_HEALTH_PINGS === 'true') {
+                return
+            }
+
             if (!this.config) {
                 throw new Error('UpstreamHealthProvider not initialized')
             }
 
-            const headers = new Headers(this.config.customHeaders as HeadersInit)
-            headers.set('Content-Type', 'application/json; charset=utf-8')
+            const sharedHeaders = new Headers(this.config.customHeaders as HeadersInit)
+            sharedHeaders.set('Content-Type', 'application/json; charset=utf-8')
+            addTraceparent(sharedHeaders)
+            addCustomUserAgent(sharedHeaders)
+
+            const upstreamHeaders = new Headers(sharedHeaders)
             if (this.config.accessToken) {
-                headers.set('Authorization', `token ${this.config.accessToken}`)
+                upstreamHeaders.set('Authorization', `token ${this.config.accessToken}`)
             }
-            addTraceparent(headers)
-            addCustomUserAgent(headers)
             const url = new URL('/healthz', this.config.serverEndpoint)
+            const upstreamResult = await wrapInActiveSpan('upstream-latency.upstream', span => {
+                span.setAttribute('sampled', true)
+                return measureLatencyToUri(upstreamHeaders, url.toString())
+            })
 
-            // We use a GET request even though we do not want to consume the
-            // body to avoid internal networks interfering with the request.
-            //
-            // To make sure the content is garbage collected, we'll ensure that
-            // the body is consumed. We don't use undici yet but that might
-            // change in the future:
-            //
-            // https://undici.nodejs.org/#/?id=garbage-collection
-            const response = await fetch(url.toString(), { method: 'GET', headers })
-            void response.arrayBuffer() // consume the body
-
-            const duration = Date.now() - start
-            this.pushDuration(duration)
-
-            logDebug(
-                'UpstreamHealth',
-                `Ping took ${Math.round(duration)}ms (Median: ${Math.round(
-                    this.getMedianDuration() ?? 0
-                )}ms)`,
-                {
-                    verbose: {
-                        duration,
-                        url,
-                        status: response.status,
-                        headers: headersToObject(response.headers),
-                    },
+            // We don't want to congest the network so we run the test serially
+            if (isDotCom(this.config.serverEndpoint)) {
+                const gatewayHeaders = new Headers(sharedHeaders)
+                const uri = 'https://cody-gateway.sourcegraph.com/-/__version'
+                const gatewayResult = await wrapInActiveSpan('upstream-latency.gateway', span => {
+                    span.setAttribute('sampled', true)
+                    return measureLatencyToUri(gatewayHeaders, uri)
+                })
+                if (!('error' in gatewayResult)) {
+                    this.lastGatewayLatency = gatewayResult.latency
                 }
-            )
+            }
+
+            if ('error' in upstreamResult) {
+                this.lastUpstreamLatency = undefined
+
+                logDebug('UpstreamHealth', 'Failed to ping upstream host', {
+                    verbose: {
+                        error: upstreamResult.error,
+                    },
+                })
+            } else {
+                this.lastUpstreamLatency = upstreamResult.latency
+
+                logDebug(
+                    'UpstreamHealth',
+                    `Ping took ${Math.round(upstreamResult.latency)}ms ${
+                        this.lastGatewayLatency
+                            ? `(Gateway: ${Math.round(this.lastGatewayLatency)}ms)`
+                            : ''
+                    }`,
+                    {
+                        verbose: {
+                            Latency: upstreamResult.latency,
+                            url,
+                            status: upstreamResult.response.status,
+                            headers: headersToObject(upstreamResult.response.headers),
+                        },
+                    }
+                )
+            }
         } catch (error) {
             // We don't care about errors here, we just want to measure the latency
         } finally {
@@ -112,22 +149,6 @@ export class UpstreamHealthProvider implements vscode.Disposable {
             clearTimeout(this.nextTimeoutId)
         }
     }
-
-    private pushDuration(duration: number): void {
-        const entry = { timestamp: Date.now(), duration }
-
-        // Delete items that are older than 2 hours
-        this.recentDurationsSorted = this.recentDurationsSorted.filter(
-            item => Date.now() - item.timestamp < TWO_HOURS_MS
-        )
-
-        const position = this.recentDurationsSorted.findIndex(item => item.duration > duration)
-        if (position === -1) {
-            this.recentDurationsSorted.push(entry)
-        } else {
-            this.recentDurationsSorted.splice(position, 0, entry)
-        }
-    }
 }
 
 export const upstreamHealthProvider = new UpstreamHealthProvider()
@@ -138,4 +159,26 @@ function headersToObject(headers: BrowserOrNodeResponse['headers']) {
         result[key] = value
     }
     return result
+}
+
+async function measureLatencyToUri(
+    headers: Headers,
+    uri: string
+): Promise<{ latency: number; response: BrowserOrNodeResponse } | { error: Error }> {
+    try {
+        const start = performance.now()
+        // We use a GET request even though we do not want to consume the
+        // body to avoid internal networks interfering with the request.
+        //
+        // To make sure the content is garbage collected, we'll ensure that
+        // the body is consumed. We don't use undici yet but that might
+        // change in the future:
+        //
+        // https://undici.nodejs.org/#/?id=garbage-collection
+        const response = await fetch(uri, { method: 'GET', headers })
+        void response.arrayBuffer() // consume the body
+        return { latency: performance.now() - start, response }
+    } catch (error: any) {
+        return { error }
+    }
 }